pyproject.toml
[tool.poetry]
name = "caixa-de-ferramentas"
version = "0.1.2"
description = ""
authors = ["Marco Menezes <marcoaurelioreislima@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
redis = "^5.0.0"
kafka-python = "^2.0.2"
azure-data-tables = "^12.4.3"
azure-keyvault-secrets = "^4.7.0"
azure-storage-blob = "^12.17.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
